Fix crashes on 6 plus and memory leak issue
A crash was occuring on the iPhone 6+ due to there being more bytes per pixel
than other models (6 bytes vs. 4). Instead of trying to figure out
which model we're running on, we leave it to Core Graphics to allocate a
context of the correct size.

Core Graphics also manages the memory that it allcoates for this
context, meaning it is freed when the context ref is released. This
solves the memory leak issue in #4.